Add multiple build types.
Allows for multiple builds based on possible needs of the end user, which each type increasing the output file size.  The test files include allowing a test to run only for an executable compiled for a specific capability.  Additionally, the new 'version' command is added to inform the user (and tests) what capabilities it includes.