Improve exec command.
Fixed out bugs in the exec command, where it wasn't properly parsing the arguments.  The tests were updated, because they weren't quite right.  The version now includes the +exec as well.

The make was expanded to include debug versions of slim and fat files, and a new extract script added to make extraction of executables from a built container easier.