Changes to version label to include library, improved signal handling in dietlibc, improved number format checking, added testing to builds.
The version number bumped from 1.1 to 1.1.1, due to bug fixes in signal handling and number format error checking.

Signal handling improved because of issues discovered with dietlibc.  dietlibc uses the older POSIX standard for sigwait() which caused incorrect error checking.  This is improved, but there's still an issue with the interesting scenario for ignoring some signals and not waiting on them.

Number format error checking is vastly improved to better detect when something is not a numeric value, or has non-numeric text after a number.  This was causing issues with glibc.

Builds for each library version now include running tests, to ensure the code is compatible with that library.  The built file also includes the library it is statically linked against in the version description.

Added tests for better exercising some of the error conditions discovered.