Version 2.0.1 wrapping up.
Wrapped the require-all-args feature in a new define (REQUIRE_FULL_CMD).  This might or might not be the standard execution mode, but will require changing the tests, which probably means a major version bump.  But this is only for error conditions, so maybe not.

Altered the make to not require Python to build (issue #4).  If Python is not installed and the headers are not generated, then the build will still fail.

Added a new test build that checks that each individual command flag can build and passes tests.  This also ensures that, with the flag enabled, the  command reports the command enabled (issue #3).

Changed the automated builds to instead run the build once for each command.  The old combinatorics test still exists, but it's been renamed and will only build on pushes to the  branch.  This will likely fail because the test takes over 6 hours.

Fixed a bug in the tests that would cause rm only command builds to also run rmdir commands.

Updated documentation to better match the current version.