Bug fixes, test improvements, and doc improvements.
Fixed an issue with allocating memory.  The type sizes were not 100% accurate, but this only led to allocating more memory than what was actually needed.

Fixed an issue with tests not running the version command tests.

Updated the  to show how to fully build the command + copy it into a scratch org from a multi-part Dockerfile.