argument parser error checking (#10)
With this, the argument parsing routine was rewritten which made the compled version smaller.  This significantly changes how non -c argument parsing is handled now, and it may not be supported for v3.0.

Fixed a bug with the exec command when the exec call out fails.  Before, the error was ignored.  Added a test to cover this (exec-not-a-command.sh).

Marked REQUIRE_FULL_CMD as a top-level makefile flag.