Add environment variable parsing support with new flag.
Lots of tests added to test out the parsing.

Tests updated to allow a new - prefix on flags to run only if the flag is not turned on.

Updated non-command compile flags to be included in the main Makefile flag file, and included them in the combinatorics builds.

Fixed an issue with the arg parser rework to report a message if the argument was too long at first check.  Also updated the error message to conform.