Migrate to loading input from a new module.
Environment variables now use the standard library setenv / getenv.  Need to check if that ends up making the code bigger than it needs to be.

Exec changed so that it takes a single argument, which is then sub-parsed into individual arguments for the command.  This is made possible mostly because of the quoting format changes.

Quoting characters to group arguments together has changed.  It now uses [ to start a quoted section and ] to end it, which allows for embedded quote tracking.  This is a significant breaking change, and makes the shell even more POSIX non-compliant.