Dev Branch stabalization.
Clean up gen-cmd with out-dated files.

Add tests for ls-l, and update the makefile to use the right ls header.  This allows the ls-s command to be built now.

Removed an extra printf that escaped debugging.

Removed difftime call, which introduced a subtle floating point dependency bloat.

Added in start-timer and elapsed-time and export-elapsed-time commands.