Fixed up tests for ls-l to be independent of platform sorting algorithms.
Added elapsed-time-under, to work well with network checking commands.

Added a build mode (interative) to allow for creating an image to work with without building anything.

Added a recipe to show off loops, timers, and http checks (delay-until-ready.Dockerfile).

Made 'ArgReplaceNewline' in the argument parser available for other commands to inspect or alter.