Fix typos and Grammar
There are typos and grammar mistakes in the readme file, and some of these were cleaned up.

With this, I noticed that one of the command flags, 'USE_ENVIROMENT_INPUT', included a typo.  I've switched that to now have a correctly spelled version, while allowing the old one to still work via a new vcompat.h file.  This requires bumping the version number up to 4.5.