Many fixes for tests related to networking; Added LLVM fixes
The LLVM implementation failed to build for the mkdir because of a missing header.

The LLVM would fail to compile the debug version on some versions, causing a crash due to conflicting debug flags.

The tests that used networking would run into issues if the host network didn't support IPv6.  The tests now recognize that, and use it as another skippable flag.

The network tests also would fail if a port was in use.  Now, the tests run a helper utility to run the test server, trying multiple times, with pseudo-random port numbers.

Test docker images updated with newer versions and with newer dependency versions.