Minor tweaks to allow IPv6 settings to pass into the samples
The test builds may run in an environment that does not support IPv6, in which case those tests must not run (or they will fail).  In even more extreme circumstances, the ip v6 detection mechanism does not work as expected.  Due to these reasons, the builds must allow external declaration on whether to run with IPv6 tests.

This becomes obviously clear when run in the GitHub actions, which don't support IPv6 from docker.