Update tests based on virtual network environments
With a virtual network environment, such as what Docker creates, the network will accept all socket connections, even if nothing listens on it, and only at a write will generate a SIGPIPE signal.  A future command will handle this strange behavior.  Until that time, the tests show how it will be used (with a 'trap' command), and splits out the tests to include one that requires a 'real' network.

Also, fixed some tests to indicate they require ipv6.