The new compressed executable method
Added in an initial cut at the new compressed executable approach (which doesn't run right), and has minimal improvement to the file size.  However, there exist some situations where it looks like it could have a smaller size, so may indicate a place for future investigation to fix.

Taught the docker build command (default operation of 'internal-docker-make.sh') to create the smaller version of the 'presh' executable as 'out/presh-smallest'.  This allows later stages in a 'Dockerfile' to use the smallest version, if it cares about that.

Updated the 'recipes' to use the smallest file size, as they care about that.

Updated the file sizes in the 'recipes' to reflect the new file sizes.

Added a few improvements to the root 'Dockerfile' to indicate its differences from the 'sample.Dockerfile', and to fix its use of 'CMD'.

Updated the license copyright, and added an SPDX to it, and added a note about static compilation affects on the binary distribution.